Split twilio.ts into separate files
